+++ cli/docs/changelog.md
@@ -1,4 +1,6 @@
 # TBD
+### Changes
+* Upgraded to container-engine-lib 0.29.1, core 1.50.4, and engine 1.22.4
 
 # 0.18.4
 ### Features