+++ cli/docs/changelog.md
@@ -2,6 +2,9 @@
 ### Changes
 * Upgraded to container-engine-lib 0.29.1, core 1.50.4, and engine 1.22.4
 
+### Features
+* `kurtosis gateway` now connects to user services running in cluster
+
 # 0.18.4
 ### Features
 * Added files artifact expansion in Kubernetes