+++ cli/docs/changelog.md
@@ -1,4 +1,6 @@
 # TBD
+
+# 0.18.5
 ### Changes
 * Upgraded to container-engine-lib 0.30.0, core 1.50.5, and engine 1.22.5